requirements.txt
pandas==2.0.0
numpy==1.24.0
scikit-learn==1.2.0
lifetimes==0.11.3
matplotlib==3.7.0
seaborn==0.12.0
pyyaml==6.0.1
pytest==7.3.1
jupyter==1.0.0
